Module_6/API/iris-classifier/src/requirements.txt
astroid==2.3.3
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
pylint==2.4.4
pyproj==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
scikit-learn==0.22.1
scipy==1.5.2
six==1.14.0
threadpoolctl==2.1.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
